Don't remove app if test failed.

+++ integration-tests/tests/basic-happy-path.spec.ts
@@ -19,7 +19,18 @@ describe('Basic Happy Path', { tags: ['@PR-check', '@publicRepo'] }, () => {
   const quarkusDeplomentBody = 'Congratulations, you have created a new Quarkus cloud application';
 
   after(function () {
-    Applications.deleteApplication(applicationName);
+    // If some test failed, don't remove the app
+    let allTestsSucceeded = true;
+    if (
+      this.test.parent.eachTest((test) => {
+        if (test.state == 'failed') {
+          allTestsSucceeded = false;
+        }
+      })
+    )
+      if (allTestsSucceeded || Cypress.env('REMOVE_APP_ON_FAIL')) {
+        Applications.deleteApplication(applicationName);
+      }
   });
 
   it('Create an Application with a component', () => {
